OCMAKU-177 partial commit due to power failure :)

+++ web/src/main/java/org/devgateway/ocds/web/convert/MakueniToOCDSConversionServiceImpl.java
@@ -220,7 +220,7 @@ public class MakueniToOCDSConversionServiceImpl implements MakueniToOCDSConversi
         Item ocdsItem = new Item();
         safeSet(ocdsItem::setId, tenderItem::getId, this::longIdToString);
         safeSet(ocdsItem::setUnit, () -> tenderItem, this::createTenderItemUnit);
-        safeSet(ocdsItem::setQuantity, tenderItem::getQuantity, Integer::doubleValue);
+        safeSet(ocdsItem::setQuantity, tenderItem::getQuantity, BigDecimal::doubleValue);
         safeSet(ocdsItem::setClassification, () -> tenderItem, this::createTenderItemClassification);
         return ocdsItem;
     }
@@ -523,7 +523,7 @@ public class MakueniToOCDSConversionServiceImpl implements MakueniToOCDSConversi
         safeSet(ocdsItem::setId, purchaseItem::getId, this::longIdToString);
         safeSet(ocdsItem::setDescription, purchaseItem::getLabel);
         safeSet(ocdsItem::setUnit, () -> purchaseItem, this::createPlanningItemUnit);
-        safeSet(ocdsItem::setQuantity, purchaseItem::getQuantity, Integer::doubleValue);
+        safeSet(ocdsItem::setQuantity, purchaseItem::getQuantity, BigDecimal::doubleValue);
         safeSet(ocdsItem::setClassification, () -> purchaseItem, this::createPlanningItemClassification);
         safeSet(ocdsItem::setTargetGroup, purchaseItem.getPlanItem()::getTargetGroup, this::categoryLabel);
         safeSet(ocdsItem::setTargetGroupValue, purchaseItem.getPlanItem()::getTargetGroupValue, this::convertAmount);
